Para la parte de las quejas y obras, ya pueden hacer registros y editar

+++ database/migrations/2022_03_07_161901_create_os_liquidadas_table.php
@@ -17,7 +17,7 @@ class CreateOsLiquidadasTable extends Migration
             $table->increments('id');
             $table->string('estado');
             $table->date('fecha');
-            $table->integer('telefono');
+            $table->bigInteger('telefono');
             $table->string('nombre_cliente');
             $table->string('name_tecnico');
             $table->string('pic_tecnico');
